// src/app/models/CoordinatorForAdd.ts
export class CoordinatorForAdd {
	coordinatorFirstName: string;
	coordinatorMiddleName?: string;
	coordinatorLastName: string;
	coordinatorSecondLastName?: string;
	coordinatorEmail: string;
	coordinatorRegionalContestName?: string;
	coordinatorRegionalContestLogoURI?: string;
	coordinatorRegionalContestURL?: string;
}